menambah fitur peminjaman buku untuk user

+++ database/migrations/2025_03_24_171054_remove_name_from_books.php
@@ -12,9 +12,12 @@ return new class extends Migration
     public function up()
     {
         Schema::table('books', function (Blueprint $table) {
-            $table->dropColumn('name');
+            if (Schema::hasColumn('books', 'name')) {
+                $table->dropColumn('name');
+            }
         });
     }
+
     
     public function down()
     {
